# public/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API SpaceCoder",
        "description": "An API for SpaceCoder web appliation.",
        "termsOfService": "https://fr.spacecoder.fun/terms-and-conditions",
        "contact": {
            "name": "SpaceCoder API Team",
            "url": "https://fr.spacecoder.fun/contact",
            "email": "contact@spacecoder.fun"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://fr.spacecoder.fun/api/",
            "description": "API SpaceCoder"
        },
        {
            "url": "http://localhost:8000/api/",
            "description": "API Local"
        }
    ],
    "paths": {
        "/articles/{slug}": {
            "get": {
                "tags": [
                    "Article"
                ],
                "summary": "Info for a specific article",
                "operationId": "article",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/slug"
                    },
                    {
                        "name": "comments",
                        "in": "query",
                        "description": "Return article detail with it's comments",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An article details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Article"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "An article details + Comments",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "article": {
                                            "$ref": "#/components/schemas/Article"
                                        },
                                        "comments": {
                                            "title": "Comments",
                                            "description": "List of Comments of article",
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Comment"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                }
            }
        },
        "/articles": {
            "get": {
                "tags": [
                    "Article"
                ],
                "summary": "List all articles",
                "description": "Une Bonne description bien longue",
                "operationId": "articles",
                "responses": {
                    "200": {
                        "description": "List of Articles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Articles"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                },
                "deprecated": false
            }
        },
...
        "/comments/{id}": {
            "get": {
                "tags": [
                    "Comment"
                ],
                "summary": "Info for a specific comment",
                "operationId": "comment",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The id of the comment",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "replies",
                        "in": "query",
                        "description": "Return comment detail with it's replies",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A comment details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Comment"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "A comment details + Replies",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "comment": {
                                            "$ref": "#/components/schemas/Comment"
                                        },
                                        "replies": {
                                            "title": "Replies",
                                            "description": "List of Replies of comment",
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Comment"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                }
            }
        },
        "/replies/{id}": {
            "get": {
                "tags": [
                    "Comment"
                ],
                "summary": "Get replies of a comment",
                "operationId": "replies",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The id of the comment",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A comment details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Comment"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                }
            }
        },
        "/comments": {
            "get": {
                "tags": [
                    "Comment"
                ],
                "summary": "List all comments",
                "operationId": "comments",
                "responses": {
                    "200": {
                        "description": "List of Articles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Comments"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                }
            }
        },
        "/article-comments/{slug}": {
            "get": {
                "tags": [
                    "Comment"
                ],
                "summary": "Get comments of an article",
                "operationId": "commentsOfArticle",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "The slug of the article",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of Article's comments",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Comments"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Articles": {
                "description": "Article Collection",
                "properties": {
                    "uri": {
                        "type": "string"
                    },
                    "id": {
                        "type": "integer"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "subject": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "views": {
                        "type": "integer"
                    },
                    "publishedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "comment": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "Article": {
                "description": "Article Item",
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/Articles"
                    },
                    {
                        "properties": {
                            "level": {
                                "type": "integer"
                            },
                            "content": {
                                "type": "string"
                            },
                            "author": {
                                "type": "string"
                            },
                            "suggeredBy": {
                                "type": "string"
                            },
                            "createAt": {
                                "type": "string",
                                "format": "date-time"
                            }
                        }
                    }
                ]
            },
            "Comments": {
                "description": "Comment Collection",
                "properties": {
                    "uri": {
                        "type": "string"
                    },
                    "id": {
                        "type": "integer"
                    },
                    "author": {
                        "type": "string"
                    },
                    "article": {
                        "type": "string"
                    },
                    "content": {
                        "type": "string"
                    },
                    "replyTo": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Comment": {
                "description": "Comment Item",
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/Comments"
                    },
                    {
                        "properties": {
                            "replies": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            },
                            "publishedAt": {
                                "type": "string",
                                "format": "date-time"
                            }
                        }
                    }
                ]
            },
...
        },
        "responses": {
            "NotFound": {
                "description": "Ressource not found.",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "code": {
                                    "type": "integer",
                                    "example": 404
                                },
                                "message": {
                                    "type": "string",
                                    "example": "Ressource not found."
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        },
        "parameters": {
            "slug": {
                "name": "slug",
                "in": "path",
                "description": "Ressource slug",
                "required": true,
                "schema": {
                    "type": "string"
                }
            }
        },
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "description": ">- Enter the token with the `Bearer: {token}`, e.g. Bearer abcde12345",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}